+++ tokio/src/executor/mod.rs
@@ -52,8 +52,6 @@ pub(crate) use self::enter::enter;
 mod global;
 pub use self::global::spawn;
 
-pub(crate) mod loom;
-
 pub mod park;
 
 #[cfg(feature = "rt-current-thread")]
