+++ tokio/src/executor/task/join.rs
@@ -1,5 +1,5 @@
-use crate::executor::loom::alloc::Track;
 use crate::executor::task::raw::RawTask;
+use crate::loom::alloc::Track;
 
 use std::fmt;
 use std::future::Future;
