+++ tokio/src/executor/thread_pool/queue/mod.rs
@@ -8,7 +8,7 @@ mod worker;
 pub(crate) use self::inject::Inject;
 pub(crate) use self::worker::Worker;
 
-use crate::executor::loom::sync::Arc;
+use crate::loom::sync::Arc;
 
 pub(crate) fn build<T: 'static>(workers: usize) -> Vec<Worker<T>> {
     let local: Vec<_> = (0..workers).map(|_| local::Queue::new()).collect();
