+++ tokio/src/executor/thread_pool/shutdown.rs
@@ -3,7 +3,7 @@
 //! Each worker holds the `Sender` half. When all the `Sender` halves are
 //! dropped, the `Receiver` receives a notification.
 
-use crate::executor::loom::sync::Arc;
+use crate::loom::sync::Arc;
 use crate::sync::oneshot;
 
 #[derive(Debug, Clone)]
