+++ tokio/src/lib.rs
@@ -98,7 +98,6 @@ pub mod io;
 #[cfg(feature = "net-driver")]
 pub mod net;
 
-#[cfg(feature = "net-driver")]
 mod loom;
 
 pub mod prelude;
