+++ tokio/src/sync/mod.rs
@@ -37,8 +37,6 @@ macro_rules! if_loom {
 mod barrier;
 pub use barrier::{Barrier, BarrierWaitResult};
 
-mod loom;
-
 pub mod mpsc;
 
 mod mutex;
